Implement booking update functionality; modify availability status to 'booked' and add booking update endpoint with role-based access control

+++ src/modules/user/user.service.ts
@@ -3,10 +3,10 @@ import { pool } from "../../config/db";
 
 
 
-const getAllUserFromDB = async()=>{
-    const result = await pool.query(`SELECT id, name, email, phone, role FROM users`);    
-    return result;
-}  
+const getAllUserFromDB = async () => {
+  const result = await pool.query(`SELECT id, name, email, phone, role FROM users`);
+  return result;
+}
 
 const getSingleUser = async (id: string) => {
   const result = await pool.query(`SELECT * FROM users WHERE id = $1`, [id]);
@@ -15,13 +15,23 @@ const getSingleUser = async (id: string) => {
 };
 
 const deleteUser = async (id: string) => {
+  const bookingCheck = await pool.query(`
+      SELECT * FROM bookings WHERE customer_id = $1 AND status = 'active'
+      `, [id])
+  if (bookingCheck.rows.length > 0) {
+    return {
+      rowCount: 0,
+      success: false,
+      message: "User cannot be deleted because they have active bookings."
+    };
+  }
   const result = await pool.query(`DELETE FROM users WHERE id = $1`, [id]);
 
   return result;
 };
 
 const updateUser = async (id: string, name?: string, email?: string, phone?: string, role?: string) => {
-  const result = await pool.query( `
+  const result = await pool.query(`
       UPDATE users
       SET 
         name = COALESCE($1, name),
@@ -37,8 +47,8 @@ const updateUser = async (id: string, name?: string, email?: string, phone?: str
 };
 
 export const userServices = {
-    getAllUserFromDB,
-    getSingleUser,
-    deleteUser,
-    updateUser
+  getAllUserFromDB,
+  getSingleUser,
+  deleteUser,
+  updateUser
 }
